backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
PyMySQL
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
